+++ cmd/pgagent/options.go
@@ -4,6 +4,5 @@ import "time"
 
 type Options struct {
 	Frequency time.Duration
-	Verbose   bool
 	Version   bool
 }