+++ cmd/pgmaven/options.go
@@ -3,6 +3,5 @@ package main
 type Options struct {
 	Command string
 	Detect  string
-	Verbose bool
 	Version bool
 }